requirements-dev.txt
asgiref==3.4.1
astroid==2.8.4
autopep8==1.6.0
backcall==0.2.0
colorama==0.4.4
decorator==5.1.0
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-linter==0.7
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycodestyle==2.8.0
pygments==2.10.0
pylint==1.5.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
typing-extensions==3.10.0.2; python_version < '3.10'
wcwidth==0.2.5
wrapt==1.13.3
